move a bunch of types into dedicated modules
