+++ hadrian.cabal
@@ -22,22 +22,27 @@ executable hadrian
                        , Builder
                        , CommandLine
                        , Context
+                       , Context.Type
                        , Environment
                        , Expression
+                       , Expression.Type
                        , Flavour
                        , GHC
                        , Hadrian.Builder
                        , Hadrian.Builder.Ar
                        , Hadrian.Builder.Sphinx
                        , Hadrian.Builder.Tar
+                       , Hadrian.Builder.Type
                        , Hadrian.Expression
                        , Hadrian.Haskell.Cabal
                        , Hadrian.Haskell.Cabal.Parse
+                       , Hadrian.Haskell.Cabal.Type
                        , Hadrian.Oracles.ArgsHash
                        , Hadrian.Oracles.DirectoryContents
                        , Hadrian.Oracles.Path
                        , Hadrian.Oracles.TextFile
                        , Hadrian.Package
+                       , Hadrian.Package.Type
                        , Hadrian.Target
                        , Hadrian.Utilities
                        , Oracles.Flag
@@ -102,10 +107,12 @@ executable hadrian
                        , Settings.Packages.RunGhc
                        , Settings.Warnings
                        , Stage
+                       , Stage.Type
                        , Target
                        , UserSettings
                        , Utilities
                        , Way
+                       , Way.Type
     default-language:    Haskell2010
     default-extensions:  DeriveFunctor
                        , DeriveGeneric
