+++ src/Context/Type.hs
@@ -1,0 +1,20 @@
+module Context.Type where
+
+import Hadrian.Package.Type
+import Stage.Type
+import Way
+
+import GHC.Generics
+import Development.Shake.Classes
+
+-- | Build context for a currently built 'Target'. We generate potentially
+-- different build rules for each 'Context'.
+data Context = Context
+    { stage   :: Stage   -- ^ Currently build Stage
+    , package :: Package -- ^ Currently build Package
+    , way     :: Way     -- ^ Currently build Way (usually 'vanilla')
+    } deriving (Eq, Generic, Show)
+
+instance Binary   Context
+instance Hashable Context
+instance NFData   Context
