+++ src/Hadrian/Haskell/Cabal/Parse.hs
@@ -20,26 +20,8 @@ import qualified Distribution.Text                      as C
 import qualified Distribution.Types.CondTree            as C
 import qualified Distribution.Verbosity                 as C
 
+import Hadrian.Haskell.Cabal.Type
 import Hadrian.Package
-
--- TODO: Use fine-grain tracking instead of tracking the whole Cabal file.
--- | Haskell package metadata extracted from a Cabal file.
-data Cabal = Cabal
-    { dependencies :: [PackageName]
-    , name         :: PackageName
-    , synopsis     :: String
-    , version      :: String
-    } deriving (Eq, Read, Show, Typeable)
-
-instance Binary Cabal where
-    put = put . show
-    get = fmap read get
-
-instance Hashable Cabal where
-    hashWithSalt salt = hashWithSalt salt . show
-
-instance NFData Cabal where
-    rnf (Cabal a b c d) = a `seq` b `seq` c `seq` d `seq` ()
 
 -- | Parse a Cabal file.
 parseCabal :: FilePath -> IO Cabal
